# flex/coordinator/gs_flex_coordinator/openapi/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: graphscope@alibaba-inc.com
    name: GraphScope
  description: |-
    This is a specification for GraphScope FLEX HTTP service based on the OpenAPI 3.0 specification. You can find out more details about
    specification at [doc](https://swagger.io/specification/v3/).

    Some useful links:
    - [GraphScope Repository](https://github.com/alibaba/GraphScope)
    - [The Source API definition for GraphScope Interactive](https://github.com/GraphScope/portal/tree/main/httpservice)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GraphScope FLEX HTTP SERVICE API
  version: 0.9.1
...
servers:
- url: /
paths:
...
  /api/v1/graph:
    get:
      description: List all graphs
      operationId: list_graphs
      responses:
        "200":
          content:
            application/json:
              example:
              - name: demo
                store_type: mutable_csr
                stored_procedures:
                  directory: plugins
                schema:
                  vertex_types:
                  - type_id: 0
                    type_name: person
                    properties:
                    - property_id: 0
                      property_name: id
                      property_type:
                        primitive_type: DT_STRING
                    - property_id: 1
                      property_name: name
                      property_type:
                        primitive_type: DT_STRING
                    primary_keys:
                    - id
                  - type_id: 1
                    type_name: software
                    properties:
                    - property_id: 0
                      property_name: id
                      property_type:
                        primitive_type: DT_STRING
                    - property_id: 1
                      property_name: name
                      property_type:
                        primitive_type: DT_STRING
                    primary_keys:
                    - id
                  edge_types:
                  - type_id: 0
                    type_name: knows
                    vertex_type_pair_relations:
                    - source_vertex: person
                      destination_vertex: person
                      relation: MANY_TO_MANY
                      x_csr_params:
                        edge_storage_strategy: ONLY_IN
                    properties:
                    - property_id: 0
                      property_name: weight
                      property_type:
                        primitive_type: DT_STRING
                  - type_id: 1
                    type_name: created
                    vertex_type_pair_relations:
                    - source_vertex: person
                      destination_vertex: software
                      relation: MANY_TO_MANY
                      x_csr_params:
                        edge_storage_strategy: ONLY_IN
                    properties:
                    - property_id: 0
                      property_name: weight
                      property_type:
                        primitive_type: DT_STRING
              schema:
                items:
                  $ref: '#/components/schemas/Graph'
                type: array
          description: Successful operation
      tags:
      - graph
      x-openapi-router-controller: gs_flex_coordinator.controllers.graph_controller
    post:
      description: Create a new graph
      operationId: create_graph
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Graph'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: successful operation
      tags:
      - graph
      x-openapi-router-controller: gs_flex_coordinator.controllers.graph_controller
  /api/v1/graph/{graph_name}:
    delete:
      description: Delete a graph by name
      operationId: delete_graph
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Successful operation
      tags:
      - graph
      x-openapi-router-controller: gs_flex_coordinator.controllers.graph_controller
...
  /api/v1/graph/{graph_name}/schema:
    get:
      description: Get graph schema by name
      operationId: get_schema
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                vertex_types:
                - type_id: 0
                  type_name: person
                  properties:
                  - property_id: 0
                    property_name: id
                    property_type:
                      primitive_type: DT_STRING
                  - property_id: 1
                    property_name: name
                    property_type:
                      primitive_type: DT_STRING
                  primary_keys:
                  - id
                - type_id: 1
                  type_name: software
                  properties:
                  - property_id: 0
                    property_name: id
                    property_type:
                      primitive_type: DT_STRING
                  - property_id: 1
                    property_name: name
                    property_type:
                      primitive_type: DT_STRING
                  primary_keys:
                  - id
                edge_types:
                - type_id: 0
                  type_name: knows
                  vertex_type_pair_relations:
                  - source_vertex: person
                    destination_vertex: person
                    relation: MANY_TO_MANY
                    x_csr_params:
                      edge_storage_strategy: ONLY_IN
                  properties:
                  - property_id: 0
                    property_name: weight
                    property_type:
                      primitive_type: DT_STRING
                - type_id: 1
                  type_name: created
                  vertex_type_pair_relations:
                  - source_vertex: person
                    destination_vertex: software
                    relation: MANY_TO_MANY
                    x_csr_params:
                      edge_storage_strategy: ONLY_IN
                  properties:
                  - property_id: 0
                    property_name: weight
                    property_type:
                      primitive_type: DT_STRING
              schema:
                $ref: '#/components/schemas/Schema'
          description: Successful operation
      tags:
      - graph
      x-openapi-router-controller: gs_flex_coordinator.controllers.graph_controller
  /api/v1/graph/{graph_name}/schema/edge_edge/{type_name}:
    delete:
      description: Delete a edge type by name
      operationId: delete_edge_type
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: type_name
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: source_vertex_type
        required: true
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: destination_vertex_type
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Successful operation
      tags:
      - graph
      x-openapi-router-controller: gs_flex_coordinator.controllers.graph_controller
  /api/v1/graph/{graph_name}/schema/edge_type:
    post:
      description: Create a edge type
      operationId: create_edge_type
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeType'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: successful operation
      tags:
      - graph
      x-openapi-router-controller: gs_flex_coordinator.controllers.graph_controller
  /api/v1/graph/{graph_name}/schema/vertex_type:
    post:
      description: Create a vertex type
      operationId: create_vertex_type
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VertexType'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: successful operation
      tags:
      - graph
      x-openapi-router-controller: gs_flex_coordinator.controllers.graph_controller
  /api/v1/graph/{graph_name}/schema/vertex_type/{type_name}:
    delete:
      description: Delete a vertex type by name
      operationId: delete_vertex_type
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: type_name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Successful operation
      tags:
      - graph
      x-openapi-router-controller: gs_flex_coordinator.controllers.graph_controller
...
components:
  schemas:
    ApiResponse:
      example: operation successful
      type: string
...
    Graph:
      example:
        schema:
          vertex_types:
          - type_name: type_name
            primary_keys:
            - primary_keys
            - primary_keys
            type_id: 0
            properties:
            - property_type:
                primitive_type: DT_DOUBLE
              property_id: 6
              property_name: property_name
            - property_type:
                primitive_type: DT_DOUBLE
              property_id: 6
              property_name: property_name
          - type_name: type_name
            primary_keys:
            - primary_keys
            - primary_keys
            type_id: 0
            properties:
            - property_type:
                primitive_type: DT_DOUBLE
              property_id: 6
              property_name: property_name
            - property_type:
                primitive_type: DT_DOUBLE
              property_id: 6
              property_name: property_name
          edge_types:
          - type_name: type_name
            type_id: 1
            vertex_type_pair_relations:
            - source_vertex: source_vertex
              destination_vertex: destination_vertex
              x_csr_params:
                edge_storage_strategy: ONLY_IN
              relation: MANY_TO_MANY
            - source_vertex: source_vertex
              destination_vertex: destination_vertex
              x_csr_params:
                edge_storage_strategy: ONLY_IN
              relation: MANY_TO_MANY
            properties:
            - property_type:
                primitive_type: DT_DOUBLE
              property_id: 6
              property_name: property_name
            - property_type:
                primitive_type: DT_DOUBLE
              property_id: 6
              property_name: property_name
          - type_name: type_name
            type_id: 1
            vertex_type_pair_relations:
            - source_vertex: source_vertex
              destination_vertex: destination_vertex
              x_csr_params:
                edge_storage_strategy: ONLY_IN
              relation: MANY_TO_MANY
            - source_vertex: source_vertex
              destination_vertex: destination_vertex
              x_csr_params:
                edge_storage_strategy: ONLY_IN
              relation: MANY_TO_MANY
            properties:
            - property_type:
                primitive_type: DT_DOUBLE
              property_id: 6
              property_name: property_name
            - property_type:
                primitive_type: DT_DOUBLE
              property_id: 6
              property_name: property_name
        stored_procedures:
          directory: plugins
        name: name
        store_type: mutable_csr
      properties:
        name:
          title: name
          type: string
        store_type:
          enum:
          - mutable_csr
          title: store_type
          type: string
        stored_procedures:
          $ref: '#/components/schemas/Graph_stored_procedures'
        schema:
          $ref: '#/components/schemas/Schema'
      title: Graph
      type: object
      x-body-name: graph
    Schema:
      example:
        vertex_types:
        - type_name: type_name
          primary_keys:
          - primary_keys
          - primary_keys
          type_id: 0
          properties:
          - property_type:
              primitive_type: DT_DOUBLE
            property_id: 6
            property_name: property_name
          - property_type:
              primitive_type: DT_DOUBLE
            property_id: 6
            property_name: property_name
        - type_name: type_name
          primary_keys:
          - primary_keys
          - primary_keys
          type_id: 0
          properties:
          - property_type:
              primitive_type: DT_DOUBLE
            property_id: 6
            property_name: property_name
          - property_type:
              primitive_type: DT_DOUBLE
            property_id: 6
            property_name: property_name
        edge_types:
        - type_name: type_name
          type_id: 1
          vertex_type_pair_relations:
          - source_vertex: source_vertex
            destination_vertex: destination_vertex
            x_csr_params:
              edge_storage_strategy: ONLY_IN
            relation: MANY_TO_MANY
          - source_vertex: source_vertex
            destination_vertex: destination_vertex
            x_csr_params:
              edge_storage_strategy: ONLY_IN
            relation: MANY_TO_MANY
          properties:
          - property_type:
              primitive_type: DT_DOUBLE
            property_id: 6
            property_name: property_name
          - property_type:
              primitive_type: DT_DOUBLE
            property_id: 6
            property_name: property_name
        - type_name: type_name
          type_id: 1
          vertex_type_pair_relations:
          - source_vertex: source_vertex
            destination_vertex: destination_vertex
            x_csr_params:
              edge_storage_strategy: ONLY_IN
            relation: MANY_TO_MANY
          - source_vertex: source_vertex
            destination_vertex: destination_vertex
            x_csr_params:
              edge_storage_strategy: ONLY_IN
            relation: MANY_TO_MANY
          properties:
          - property_type:
              primitive_type: DT_DOUBLE
            property_id: 6
            property_name: property_name
          - property_type:
              primitive_type: DT_DOUBLE
            property_id: 6
            property_name: property_name
      properties:
        vertex_types:
          items:
            $ref: '#/components/schemas/VertexType'
          title: vertex_types
          type: array
        edge_types:
          items:
            $ref: '#/components/schemas/EdgeType'
          title: edge_types
          type: array
      title: Schema
      type: object
      x-body-name: schema
    VertexType:
      example:
        type_name: type_name
        primary_keys:
        - primary_keys
        - primary_keys
        type_id: 0
        properties:
        - property_type:
            primitive_type: DT_DOUBLE
          property_id: 6
          property_name: property_name
        - property_type:
            primitive_type: DT_DOUBLE
          property_id: 6
          property_name: property_name
      properties:
        type_id:
          format: int32
          title: type_id
          type: integer
        type_name:
          title: type_name
          type: string
        properties:
          items:
            $ref: '#/components/schemas/Property'
          title: properties
          type: array
        primary_keys:
          items:
            type: string
          title: primary_keys
          type: array
      title: VertexType
      type: object
      x-body-name: vertex_type
    EdgeType:
      example:
        type_name: type_name
        type_id: 1
        vertex_type_pair_relations:
        - source_vertex: source_vertex
          destination_vertex: destination_vertex
          x_csr_params:
            edge_storage_strategy: ONLY_IN
          relation: MANY_TO_MANY
        - source_vertex: source_vertex
          destination_vertex: destination_vertex
          x_csr_params:
            edge_storage_strategy: ONLY_IN
          relation: MANY_TO_MANY
        properties:
        - property_type:
            primitive_type: DT_DOUBLE
          property_id: 6
          property_name: property_name
        - property_type:
            primitive_type: DT_DOUBLE
          property_id: 6
          property_name: property_name
      properties:
        type_id:
          format: int32
          title: type_id
          type: integer
        type_name:
          title: type_name
          type: string
        vertex_type_pair_relations:
          items:
            $ref: '#/components/schemas/EdgeType_vertex_type_pair_relations_inner'
          title: vertex_type_pair_relations
          type: array
        properties:
          items:
            $ref: '#/components/schemas/Property'
          title: properties
          type: array
      title: EdgeType
      type: object
      x-body-name: edge_type
    Property:
      example:
        property_type:
          primitive_type: DT_DOUBLE
        property_id: 6
        property_name: property_name
      properties:
        property_id:
          format: int32
          title: property_id
          type: integer
        property_name:
          title: property_name
          type: string
        property_type:
          $ref: '#/components/schemas/Property_property_type'
      title: Property
      type: object
      x-body-name: property
...
    Graph_stored_procedures:
      example:
        directory: plugins
      properties:
        directory:
          enum:
          - plugins
          title: directory
          type: string
      title: Graph_stored_procedures
      type: object
    EdgeType_vertex_type_pair_relations_inner_x_csr_params:
      description: Used for storage optimization
      example:
        edge_storage_strategy: ONLY_IN
      properties:
        edge_storage_strategy:
          enum:
          - ONLY_IN
          - ONLY_OUT
          - BOTH_OUT_IN
          title: edge_storage_strategy
          type: string
      title: EdgeType_vertex_type_pair_relations_inner_x_csr_params
      type: object
    EdgeType_vertex_type_pair_relations_inner:
      example:
        source_vertex: source_vertex
        destination_vertex: destination_vertex
        x_csr_params:
          edge_storage_strategy: ONLY_IN
        relation: MANY_TO_MANY
      properties:
        source_vertex:
          title: source_vertex
          type: string
        destination_vertex:
          title: destination_vertex
          type: string
        relation:
          enum:
          - MANY_TO_MANY
          - ONE_TO_MANY
          - MANY_TO_ONE
          - ONE_TO_ONE
          title: relation
          type: string
        x_csr_params:
          $ref: '#/components/schemas/EdgeType_vertex_type_pair_relations_inner_x_csr_params'
      title: EdgeType_vertex_type_pair_relations_inner
      type: object
    Property_property_type:
      example:
        primitive_type: DT_DOUBLE
      properties:
        primitive_type:
          enum:
          - DT_DOUBLE
          - DT_STRING
          - DT_SIGNED_INT32
          - DT_SIGNED_INT64
          - DT_DATE32
          title: primitive_type
          type: string
      title: Property_property_type
      type: object
...